[nrfconnect] Improve SoftwareUpdateCompleted boot reason detection
The existing code relies on a specific bootloader mode
to detect that the last reboot was due to software update.
This method will not work correctly if a user chooses
a different bootloader mode, or clears the "tentative"
flag in the current firmware image before the diagnostic
data provider is initialized.

Store the software reboot reason explicitly in one of the
retention registers on nRF SOCs.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>