[system] Increase number of timers to 32
The maximum number of timers is set to 32, but most
platforms override this value to 16 (well, for platforms
that use heap-allocated pools, such as Darwin or Linux,
it doesn't matter).

The problem is that 16 timers is too little to handle
the spec-mandated number of active subscriptions and other
scheduled activities. Double the number of timers to pass
stress/capacity tests.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>