TC-RR-1.1: Add optional local session ID checking
This adds comparison of the local session ID for each CASE session
established at the onset of the test against that retrieved at the end
of the test to ensure a new one has not been established. This validates
step 8 of TC-RR-1.1.

This is controlled using the 'check_local_session_id_unchanged' flag.