Don't notify OnSessionEstablishmentError after OnSessionEstablished.
A PairingSession that has sent an OnSessionEstablished notification should not
send an OnSessionEstablishmentError notification after that (e.g. if the session
gets evicted).  The session is established at that point, and the
_establishment_ cannot hit an error.

This is needed to allow PASE sessions to be sanely evicted (e.g. when we're done
with them) without triggering spurious errors.