Remove KeyError encountered on unknown attributes
This removes the KeyError exception that is thrown when the Python
cluster cache encounters an attribute that it doesn't have schema for.
This occurs for some of the extended attributes in the test cluster that
are not actually defined in schema.

This ensures that we don't abruptly terminate attribute processing when
running tests against the all clusters app.