Add APIs to reroute SDK logs on Darwin platforms
[Feature]

We need to collect verbose logs to assist in debugging and root causing issues
that customers are experiencing.

[Solution]

Added a vanilla C bridge that can be used to reroute the logs from the C++ SDK
on all platforms. Added public APIs to the Darwin Framework that allows
rerouting the Darwin-specific logs and the C++ SDK logs via the C bridge.

[Testing]

A local build succeeds for My Mac (Intel x86 64), Any iPhone, and the iPhone 11
simulator. Built the Framework into an iOS application and verified that
there are no crashes during Matter commisioning / provisioning and that the logs
are rerouted as expected.