Fix in-place save for DefaultSessionResumptionStorage
Nodes that already exist in the DefaultSessionResumptionStorage index
are consuming additional entries in the index on save, and aren't
cleaned from the link and state tables first.  This has the following
consequences:

1. If session resumption storage is full, this causes unnecessary
   eviction of records for other nodes.
2. This pulls the index out of sync from the backing state table
   because duplicate entries in the index for a given node only map
   map to a single node-id-keyed entry in the state table.
3. An entry in the link table is leaked on each occurrence, as the
   resumption-id-keyed link entry will be orphaned once its associated
   state table entry is overwritten.

This commit fixes the problem by first deleting the link table entry
before saving a record for a node that already exists in the table.
Save is also simplified in this case by not rewriting the index, as
the node is already in the index.