Update DNS-SD advertisements on fail-safe timeout
When we remove a fabric during a pending add with fail-safe timeout, we
aren't updating DNS-SD adevertisements.  In addition to the obvious
consequence of continuing to advertise a now-stale identity, this can
also leave the advertisement slots full if the failed fabric addition
was to the Nth slot in an N-fabric device.

This then breaks advertisement when recommissioning to the Nth slot
thereafter for builds that do not set this:

CHIP_CONFIG_MINMDNS_DYNAMIC_OPERATIONAL_RESPONDER_LIST

For those builds, this condition persists until reboot or some other
action triggers DnssdServer::StartServer.

This commit fixes the problem by moving the DnssdServer::StartServer
call from the RemoveFabric invoke handler to the server's
OnFabricRemoved callback.  That will result in server restart both
for RemoveFabric invocations and fail-safe expiry.

Fixes https://github.com/project-chip/connectedhomeip/issues/23169