Fix delivery of urgent events to actually work correctly.
If an urgent event was emitted at a point when the ReadHandler subscribing for
it had already gotten its "min interval has elapsed" callback, we would just
mark the read handler dirty but not schedule a run of the reporting engine.
This would cause us to not report the event until something _did_ trigger such a
run (either the max interval being reached, or some other reading/reporting
activity).

The fix is to make sure ReadHandler always schedules a run when IsReportable()
becomes true.