Compile time code generation.
Picked PluginApplicationCallbacks because it is simple and added
it to a compile-time generated list.

The main exercise for this is to make sure we can inject the
appropriate calls to compile time generation to all examples.
It is not yet perfect as it only involves include directories,
however it does a "if it compiles it works, otherwise it fails
to compile" approach.

Uses codegen.py rather than ZAP for this particular file. For
other files, we may chose to use ZAP however for that we have
to:
  - ensure it is stand-alone runnable (likely as an installable app)
  - figure out any multi-processing conflicts as cmake/gn WILL run
    build steps in parallel and zap shares a common DB.