[Linux] Run only one GLib main event loop (#23320)
* Unregister from BlueZ when shutting down BLE manager

* Return CHIP_ERROR from all BlueZ helper functions

* Remove not used class method

* Remove D-Bus connection handler from PlatformManager

The D-Bus connection handler stored in the PlatformManager is not
used anywhere. It's not required to pre-connect to D-Bus, because
the connection (for given bus type) returned by the g_bus_get_sync()
function is shared with other callers.

* Run only one GLib main event loop per Matter SDK

It is not necessary to start new GLib main event loop for every new
D-Bus communication with external service. Single main event loop can
handle all such communication.

* Run WiFi IP address change listener in GLib main loop

* Properly release cancellable object on cleanup

* Explicitly wait for glib main thread to exit

* Workaround for TSAN false positive reports with glib

TSAN thinks that memory accessed before the call to g_source_attach()
(which is internally used by e.g. g_main_context_invoke() and
g_idle_add()) needs to be guarded by a mutex, otherwise that's a race
condition between the thread that is creating shared data (the current
thread) and the glib main event loop thread. In fact such race condition
does not occur because g_source_attach() acts here as pthread_create() -
code is not executed on the other thread before the g_source_attach() is
called.

Co-authored-by: Andrei Litvin <andy314@gmail.com>