Moving zap templates away from zap helpers which use State (eg. ensureClusters, etc) and other cleanup
- Switching to zcl_attributes_server from chip_server_cluster_attributes
- Using if_unsupported_attribute_type instead of  unless (isStrEqual chipCallback.name Unsupported)
- Using if_attribute_complex instead of if_basic_global_response
- Updating the zap repo
- Logic behind the diff in the applications: Some templates have the following hierarchy
{{#chip_client_clusters}}
{{#chip_server_cluster_attributes}}
In here we are going through the enabled client clusters and then the enabled server attributes for that cluster. However chip_server_cluster_attributes does not check if the server side cluster is actually enabled or not. The way zap is designed is to store user settings so a user may enable server cluster, enable server attribute and then disable server cluster. The backend saves user's enabled server attributes such that if in the future the user enables the server side cluster again then they get their old selections back. However that does not mean the server side attribute is enabled.
Switching to the following:
{#chip_client_clusters}}
{{#zcl_attributes_server}}
Gets all available server side attributes for the cluster. I believe this will give a better and consistent representation and ease of use for a user building the app. App developers do not need to worry if they enabled the server side cluster and attribute or not. Also they do not have to enable the server side attribute if not needed.
- Github: ZAP#682