[Zephyr] Allow to erase entire NVS on factory reset (#23434)
* [Zephyr] Allow to erase entire NVS on factory reset

Currently, the default factory reset implementation simply
removes Matter (and OpenThread) settings instead of
formatting the non-volatile storage partition. The latter
approach may be advisable in the cases where no settings
need to survive a factory reset, as it is expected to work
properly regardless of the state of the storage partition.
Additionally, it will regain the original storage
performance if it has been loaded with a lot of stale data.

Signed-off-by: Damian Krolik <damian.krolik@nordicsemi.no>