[ESP32] Make all parameters optional in partition generator script
Initially script was a bit rigid and few parameters were mandatory and
its a bit cumbersome to test if you just need to test with different
discriminator.

Added an option --no-bin which will only generate the nvs partition csv
file and not the nvs partition binary. This is helpful when matter
specific data is needed and not the actual binary.