Make sure we don't use the Matter dispatch queue when it's not running.
We've had issues where things get queued to the Matter dispatch queue while it's
not running, and then when we start it up again those things run too early in
startup and break in interesting ways.

The fix is to make sure we only queue things to the Matter dispatch queue when
they're associated with a currently-running controller.

Fixes https://github.com/project-chip/connectedhomeip/issues/22847