Make queue handling consistent for commands in MTRDevice/MTRClusters.
invokeCommandWithEndpointID on MTRDevice ran the async work item management
logic on the MTRDevice's work queue, and queued an async callback to the client
queue for the command response.  This was different from what MTRClusters did:
that ran the management logic on the client queue as well.

This PR aligns on the invokeCommandWithEndpointID behavior, because the
MTRClusters behavior allows a poorly-behaved client that blocks inside the
completion handler to completely block the device's async callback queue
processing.