Fix code to enable dmalloc & disable it (#23977)
Currently dmalloc is disabled because of a stray print from GN combined
with a strict comparison of GN's output in configuring this checker.
This caused tests to timeout and fail in a previous attempt to remove
the print in #23687.

This checker is extremely expensive, especially for tests with more
allocations. For one test, it increases the cost from 1 second to 1
hour:

% time ./out/tests/TestCommissionerDUTVectors
./out/tests/TestCommissionerDUTVectors  3325.35s user 0.05s system 99% cpu 55:25.65 total

% unset DMALLOC_OPTIONS
% time ./out/tests/TestCommissionerDUTVectors
./out/tests/TestCommissionerDUTVectors  1.24s user 0.01s system 99% cpu 1.249 total

This is much too costly to be enabled for all unit tests in presubmit,
so disable it. Other tools such as ASAN have much lower overhead.