doc: add sphinx documentation framework
Added the basics for a sphinx-based doc portal.

Added an extension `external_content` that copies the relevant
doc files that are used in the documentation. This extension
also modifies files while copying to adjust links and allow
syntax that renders on both GitHub and Sphinx.

Warnings when building the documentation have been
fixed or removed. Fixes include:
- Linking errors
- Typos
- ToC tree errors
- Adding MyST configuration data
- Removing anchors on the form <a name="my-anchor"></a> as they
  cannot be used by sphinx
- Changed the GitHub rendered emoji `:heavy_check_mark:` to the
ascii symbol ✔ so that it renders with Sphinx

Signed-off-by: Gaute Svanes Lunde <gaute.lunde@nordicsemi.no>