Move Level Control OnOff feature check into OnOff cluster. (#23820)
This simplifies the coupling a bit by doing this runtime feature check in the
same place where we are already doing the runtime check for existence of a
relevant Level Control cluster.  This allows us to remove some code in Level
Control that was duplicating existing code in On/Off.