[tlv] Optimize code size of processing context tags
TLV tag is represented as uint64_t that encodes the
following components:
- 16-bit vendor ID
- 16-bit profile number
- 32-bit tag number

Context tags, which account for vast majority of tag usage
in the SDK, are encoded as having both vendor ID and profile
number equal to 0xFFFF. Anonymous tags are encoded in the
same way, but using 0xFFFFFFFF tag number.
This is correct because vendor IDs higher than 0xFFF0 shall
not be assigned to real manufacturers, but constructing
0xFF... constants in hundreds of places adds non-negligible
overhead to the flash usage.

Encode profile ID in the negated form internally to optimize
the code size when using special tags.