[crypto] Added Method That Finds and Replaces Resigned Version of a Certificate.
This method checks for resigned version of the reference certificate in the list and returns it.

The following conditions SHOULD be satisfied for the certificate to qualify as
a resigned version of a reference certificate:
  - SKID of the candidate and the reference certificate should match.
  - SubjectDN of the candidate and the reference certificate should match.

There is no specific use case for this method in Matter. However, specific ecosystem
implementations may find this method useful.

Some of the potential use cases could be finding resigned version of a PAI or DAC certificate.

Also, this method can be useful when Matter introduces attestation certificate revocation mechanism.