[v1.0 cherrypick] Update LevelControl to allow for disabling transition time and OnOff feature (#24187)
* [Level Control] Added ability to disable transitioning and OnOff feature (#23679)

* Added ability to disable transitioning and onoff feature

* Fixes from restyle

* Fixed missing OnOff attribute change

* Only handle OnOff state for off

* Move Level Control OnOff feature check into OnOff cluster. (#23820)

This simplifies the coupling a bit by doing this runtime feature check in the
same place where we are already doing the runtime check for existence of a
relevant Level Control cluster.  This allows us to remove some code in Level
Control that was duplicating existing code in On/Off.