Move weakly typed enum list back into Matter repository. (#24270)
Instead of hardcoding this list in the helper (in the ZAP repository),
it should be data in the Matter repository.  Then we can remove things
from the list without having to synchronize with ZAP changes.

This change on its own does not make ZAP use this YAML instead of the
hardcoded list; that will need to be a separate ZAP change after this
merges.