doc: move link targets to build folder
Created a change in the `external_content` extension so that
links to content that is also being copied to the _build folder
will point to their new location instead. This enables linking
to sections on other pages.

This change also enables link checking for sections on other
pages, and therefore also includes a fix to some previously
broken section links.

Signed-off-by: Gaute Svanes Lunde <gaute.lunde@nordicsemi.no>