Fix syntax of test in chip_xcode_build_connector.sh.
There was a stray ']' that was being treated as part of the string, so we never
set target_arch or target_cpu.

Fixing that ended up with us setting target_arch/cpu to arm64 if it came before
x84_64 in ARCHS, which breaks CI.  So make sure we use the host arch as the
target_arch if it's in ARCHS at all.