Fix reference to out-of-scope temporary in SetUpCodePairer.
The old code used to get a reference to a member of a temporary, but nothing
kept the temporary alive (it was not assigned to a named reference or anything),
so it went out of scope and got destroyed.  And then we were referencing
garbage.

Explicitly keep references to all the things we are using, so they don't go away
on us.