workflows: add docbuild to PRs
Added the docbuild workflow to PRs to inform contributors
whether any changes will break the documentation build.

Also limited the deploy step to only run on pushes to the
master branch.

Added the `--keep-going` flag to the Sphinx makefile to list all
errors instead of stopping at the first one.

Signed-off-by: Gaute Svanes Lunde <gaute.lunde@nordicsemi.no>