Fix a startup crash when a node wants to advertise as a commissioner but not as commissionable.
In DnssdServer::Advertise we were trying to get things like a discriminator and
other commissionee information even if we were just advertising as a
commissioner (commissionableNode was false).  This could lead to crashes due to
calling things like GetCommissionableDataProvider() when we are not in fact
commissionable and don't have one.

The fix is to reorder things a bit so we set up all the advertiseParameters
state that both commissioners and commissionable share, and then condition all
the commissionable-only bits on commissionableNode.