[dnssd] Replace invalid forced reset logic
The platform implementation of DNS-SD contains a mechanism
for restarting the advertising when the platform returns
the CHIP_ERROR_FORCED_RESET error. This mechanism is broken
as it assumes that only one operational service is used and
it uses extra RAM although it is only used by Linux.

Switch to an event-based approach that allows the DNS-SD
server to restart the advertising when needed.