Call stopControllerFactory atexit in the Darwin framework.
We have lots of things with static destructors that assert clean
shutdown in the destructor.  That means that calling exit() without a
clean shutdown is pretty much guaranteed to lead to a shutdown crash.

Just ensure that we are in fact shutting down the
MTRDeviceControllerFactory if exit() is called, to avoid those crashes.

Fixes https://github.com/project-chip/connectedhomeip/issues/14859