[K32W0] Add recovery mechanism for factory data
If a power loss occurs during factory data update through OTA, the board
might be in a state in which it has erased factory data section, causing
the app task to fail at initialization.

During initialization, the factory data provider should check if a certain
predefined PDM id exists. If it does, then most likely the internal flash
section for factory data is empty, so a restore from external flash must
be done.

Signed-off-by: Marius Tache <marius.tache@nxp.com>